requirements.txt
# =============================================================================
# Core Framework
# =============================================================================
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1

# =============================================================================
# Database
# =============================================================================
asyncpg==0.30.0

# =============================================================================
# Authentication
# =============================================================================
python-jose[cryptography]==3.4.0
httpx==0.28.1

# =============================================================================
# OpenTelemetry
# =============================================================================
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv==1.1.0
